pyproject.toml
[project]
name = "tripforge"
version = "0.1.0"
description = "AI Holiday Planner — Greek Islands MVP"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.42.0",
    "httpx>=0.27.0",
    "playwright>=1.49.0",
    "jinja2>=3.1.4",
    "pydantic>=2.10.0",
    "diskcache>=5.6.0",
    "typer>=0.15.0",
    "pyyaml>=6.0",
    "rich>=13.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
tripforge = "planner:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
